fix structure pre scala3
