--- binlog-stream-models/app/io/laserdisc/mysql/binlog/models/EventHeaderV4.scala
@@ -1,8 +1,0 @@
-package io.laserdisc.mysql.binlog.models
-
-import com.github.shyiko.mysql.binlog.event.{EventHeaderV4 => JEventHeaderV4, EventType}
-
-object EventHeaderV4 {
-  def unapply(arg: JEventHeaderV4): Option[(EventType, Long, Long)] =
-    Some((arg.getEventType, arg.getTimestamp, arg.getNextPosition))
-}
