--- binlog-stream-models/app/io/laserdisc/mysql/binlog/models/RotateEventData.scala
@@ -1,7 +1,0 @@
-package io.laserdisc.mysql.binlog.models
-
-import com.github.shyiko.mysql.binlog.event.{RotateEventData => JRotateEventData}
-object RotateEventData {
-  def unapply(arg: JRotateEventData): Option[(String, Long)] =
-    Some((arg.getBinlogFilename, arg.getBinlogPosition))
-}
