--- binlog-stream-models/app/io/laserdisc/mysql/binlog/models/TableMapEventData.scala
@@ -1,8 +1,0 @@
-package io.laserdisc.mysql.binlog.models
-
-import com.github.shyiko.mysql.binlog.event.{TableMapEventData => JTableMapEventData}
-
-object TableMapEventData {
-  def unapply(arg: JTableMapEventData): Option[(Long, String, String)] =
-    Some((arg.getTableId, arg.getDatabase, arg.getTable))
-}
