--- binlog-stream-models/app/io/laserdisc/mysql/binlog/models/XidEventData.scala
@@ -1,7 +1,0 @@
-package io.laserdisc.mysql.binlog.models
-
-import com.github.shyiko.mysql.binlog.event.{XidEventData => JXidEventData}
-
-object XidEventData {
-  def unapply(arg: JXidEventData): Option[Long] = Some(arg.getXid)
-}
