+++ binlog-stream-models/src/main/scala/io/laserdisc/mysql/binlog/models/DeleteRowsEventData.scala
@@ -1,0 +1,9 @@
+package io.laserdisc.mysql.binlog.models
+
+import com.github.shyiko.mysql.binlog.event.DeleteRowsEventData as JDeleteRowsEventData
+
+import scala.jdk.CollectionConverters.*
+object DeleteRowsEventData {
+  def unapply(arg: JDeleteRowsEventData): Option[(Long, List[Array[Serializable]], Array[Int])] =
+    Some((arg.getTableId, arg.getRows.asScala.toList, arg.getIncludedColumns.stream().toArray))
+}
