+++ binlog-stream-models/src/main/scala/io/laserdisc/mysql/binlog/models/QueryEventData.scala
@@ -1,0 +1,20 @@
+package io.laserdisc.mysql.binlog.models
+
+import com.github.shyiko.mysql.binlog.event.QueryEventData as JQueryEventData
+
+object QueryEventData {
+  def truncateTable(sql: String) = {
+    val regex = "truncate( table)? (.+)".r
+    sql match {
+      case regex(_, table) => Some(table)
+      case _               => None
+    }
+  }
+  def unapply(arg: JQueryEventData): Option[(String, String, String, Option[String])] =
+    truncateTable(arg.getSql.toLowerCase) match {
+      case Some(table) =>
+        Some((arg.getSql.toLowerCase, arg.getDatabase, "truncate", Some(table)))
+      case None =>
+        Some((arg.getSql.toLowerCase, arg.getDatabase, "query", None))
+    }
+}
