+++ binlog-stream-models/src/main/scala/io/laserdisc/mysql/binlog/models/RotateEventData.scala
@@ -1,0 +1,8 @@
+package io.laserdisc.mysql.binlog.models
+
+import com.github.shyiko.mysql.binlog.event.RotateEventData as JRotateEventData
+
+object RotateEventData {
+  def unapply(arg: JRotateEventData): Option[(String, Long)] =
+    Some((arg.getBinlogFilename, arg.getBinlogPosition))
+}
