+++ binlog-stream-models/src/main/scala/io/laserdisc/mysql/binlog/models/TableMapEventData.scala
@@ -1,0 +1,8 @@
+package io.laserdisc.mysql.binlog.models
+
+import com.github.shyiko.mysql.binlog.event.TableMapEventData as JTableMapEventData
+
+object TableMapEventData {
+  def unapply(arg: JTableMapEventData): Option[(Long, String, String)] =
+    Some((arg.getTableId, arg.getDatabase, arg.getTable))
+}
