+++ binlog-stream-models/src/main/scala/io/laserdisc/mysql/binlog/models/UpdateRowsEventData.scala
@@ -1,0 +1,15 @@
+package io.laserdisc.mysql.binlog.models
+
+import com.github.shyiko.mysql.binlog.event.UpdateRowsEventData as JUpdateRowsEventData
+
+import scala.jdk.CollectionConverters.*
+
+object UpdateRowsEventData {
+  def unapply(
+      arg: JUpdateRowsEventData
+  ): Some[(Long, List[(Array[Serializable], Array[Serializable])], Array[Int])] = {
+    val beforeAfter =
+      arg.getRows.asScala.toList.map(entry => entry.getKey -> entry.getValue)
+    Some((arg.getTableId, beforeAfter, arg.getIncludedColumns.stream().toArray))
+  }
+}
