+++ binlog-stream-models/src/test/scala/io/laserdisc/mysql/binlog/models/SchemaMetadataTest.scala
