+++ binlog-stream/src/main/scala/io/laserdisc/mysql/binlog/checkpoint/BinlogOffset.scala
