+++ binlog-stream/src/main/scala/io/laserdisc/mysql/binlog/config/package.scala
