+++ binlog-stream/src/test/scala/io/laserdisc/mysql/binlog/compaction/CompactTest.scala
