+++ binlog-stream/src/test/scala/io/laserdisc/mysql/binlog/stream/Sku.scala
