+++ mysql-binlog-stream-examples/src/main/scala/main/BinLogListener.scala
