+++ mysql-binlog-stream-shared/src/main/scala/io/laserdisc/mysql/binlog/event/Offset.scala
