+++ mysql-binlog-stream-shared/src/test/resources/init.sql
