+++ mysql-binlog-stream-shared/src/test/scala/io/laserdisc/mysql/binlog/config/BinLogConfigSpec.scala
